Refactor: Move parse site with clients into a use-case
